powerstructure fix

+++ WitchPack/Assets/Scripts/UI/GameUI/HeroSelectionUI/HeroSelectionUI.cs
@@ -33,7 +33,7 @@ public class HeroSelectionUI : MonoSingleton<HeroSelectionUI> , IPointerEnterHan
         gameObject.SetActive(true);
     }
 
-    public void UpdateStatBlocks(StatType shamanStatType, float newValue) => statBlockPanel.UpdateStatBlocks(shamanStatType, newValue);
+    public void UpdateStatBlocks(StatType shamanStatType, int newValue) => statBlockPanel.UpdateStatBlocks(shamanStatType, newValue);
 
     public void Hide()
     {
