+++ WitchPack/Assets/Scripts/UI/GameUI/StatEffectPopup/StatEffectPopupManager.cs
@@ -13,7 +13,7 @@ public class StatEffectPopupManager : MonoBehaviour
         _statEffectPopupHandler = Instantiate(_statEffectPopupHandlerPrefab, _parentHolder);
     }
 
-    public static void ShowPopupWindows(int EntityId, string statName, float value, bool isPercent, Color color)
+    public static void ShowPopupWindows(int EntityId, string statName, int value, bool isPercent, Color color)
     {
         _statEffectPopupHandler.ShowPopupWindows(EntityId, statName, value, isPercent, color);
     }
